PfP Submit Redo + New Pages
When making a profile for the first time now, the form will now accept and display an uploaded image but not submit it to the database until the save button is clicked.  This eventually will require the other fields to be filled out as well.
Added: About Page
Added: Contact Page